Add test

+++ domain/model/mac.go
@@ -9,7 +9,7 @@ type Mac struct {
 	ID          uint      `gorm:"primarykey"`
 	CreatedAt   time.Time `json:"created_at"`
 	Name        string    `gorm:"size:255; not null" json:"name"`
-	Inch        float64   `gorm:"not null" json:"inch"`
+	Inch        float32   `gorm:"not null" json:"inch"`
 	CPU         string    `gorm:"size:50; not null" json:"cpu"`
 	Memory      int       `gorm:"not null" json:"memory"`
 	Strage      string    `gorm:"size:20; not null" json:"strage"`
